refactor: rename templates.ts to worker-prompts.ts
More descriptive name for the file containing prompt construction and
webhook event templating for the worker loop.

Closes #684
